Bump spdlog for gcc 13 and surface CI compiler errors
The Linux CI job failed to compile. spdlog 1.11.0 bundles fmt 9.1.0, which
does not build on gcc 13 because fmt relied on a transitive <cstdint> that
gcc 13 no longer provides. ubuntu-24.04 ships gcc 13, so the job could
never have compiled. 1.15.3 bundles fmt 11.

That bump then exposed a second problem on MSVC as well: fmt 10 and newer
refuse to format an enum implicitly, and dumpInfo() passed
FSEQFile::CompressionType straight into a %d. Now cast explicitly.

CI could not tell me any of this. Compiler output only exists in the run
log, which needs repo admin rights to download, so a failed build reported
nothing beyond "exit code 2". Both jobs now tee the build and echo the
first errors as ::error:: workflow commands, which show up as annotations
on the run, and the Linux job uploads its full build log as an artifact
when it fails.

Verified the Windows Release build compiles against spdlog 1.15.3 and that
the FSEQ logging shim, which formats through fmt::sprintf, still prints
correctly while playing a sequence.

+++ src/fseq/FSEQFile.cpp
@@ -1638,7 +1638,9 @@ void V2FSEQFile::dumpInfo(bool indent) {
     }
 
     LogDebug(VB_SEQUENCE, "%sSequence File Information\n", ind);
-    LogDebug(VB_SEQUENCE, "%scompressionType       : %d\n", ind, m_compressionType);
+    // fmt 10 and newer refuse to format a scoped-ish enum implicitly, and this
+    // is a %d anyway, so make the conversion explicit.
+    LogDebug(VB_SEQUENCE, "%scompressionType       : %d\n", ind, (int)m_compressionType);
     LogDebug(VB_SEQUENCE, "%snumBlocks             : %d\n", ind, m_handler->computeMaxBlocks());
     // Commented out to declutter the logs ... we can add it back in if we start seeing issues
     //for (auto &a : m_frameOffsets) {
